mathml/relations/html5-tree/css-inline-style-interface.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MathML 'ElementCSSInlineStyle` Mixin Tests</title>
    <link
      rel="help"
      href="https://mathml-refresh.github.io/mathml-core/#dom-and-javascript"
    />
    <style>
        mi { background-color: red;  }
      }
    </style>
    <meta name="assert" content="a thing." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

  </head>
  <body tabindex="-1">
    <span tabindex="-1">This tests the presence and functionality of features of the `ElementCSSInlineStyle` interface for MathMLElements</span>
    <math
      tabindex="-1"
      xmlns="http://www.w3.org/1998/Math/MathML"
      display="block"
    >
      <mi>E</mi>
    </math>
  </body>
  <script>

    // spot check the functionality of several interfaces
    let el = document.querySelector("mi");
    let mathEl = document.querySelector("math");


    test(function() {
      el.style.backgroundColor = "blue";
      assert_equals(
        getComputedStyle(el).backgroundColor,
        "rgb(0, 0, 255)",
        "The style attribute should be present and be functional."
      );
    }, "The style attribute be present and be functional.");


  </script>
</html>
